+++ config/config.yml
@@ -1,28 +1,28 @@
 vm-config:
-  ova-dir: /home/vlad/frondends
+  ova-dir: /home/vlad/front
 #  ova-test: /Users/rvm/Downloads/AAUJOB/NAP/2021/defatt/virtual/vbox/
-wireguard-service:
-  endpoint: localhost
-  port: 5353
-  auth-key: deneme
-  sign-key: test
-  client-conf-dir: /etc/wireguard/
-  tls:
-    enabled: false
-    certfile: <absolute-path-to-wg-certFile>
-    certkey: <absolute-path-to-wg-certKey>
-    cafile: <absolute-path-to-wg-caFile>
+#wireguard-service:
+#  endpoint: localhost
+#  port: 5353
+#  auth-key: deneme
+#  sign-key: test
+#  client-conf-dir: /etc/wireguard/
+#  tls:
+#    enabled: false
+#    certfile: <absolute-path-to-wg-certFile>
+#    certkey: <absolute-path-to-wg-certKey>
+#    cafile: <absolute-path-to-wg-caFile>
 
-defat-config:
-  endpoint: localhost
-  port: 5454
-  tls:
-    enabled: false
-    certFile:
-    certKey:
-    caFile:
-  sign-key:
-  users-file: /Users/rvm/Downloads/AAUJOB/NAP/2021/defatt/config/users.yml
+#defat-config:
+#  endpoint: localhost
+#  port: 5454
+#  tls:
+#    enabled: false
+#    certFile:
+#    certKey:
+#    caFile:
+#  sign-key:
+#  users-file: /Users/rvm/Downloads/AAUJOB/NAP/2021/defatt/config/users.yml
 
 docker-repositories:
   - username: hknd
