+++ scenarios/scenario.yml
@@ -1,14 +1,15 @@
-name: "Test scenario"
-topic: finance
-FQDN: myfinance.cu
-story-red: "Financially motivated adversary groups executing ransomware attacks have rightfully gotten our attention in recent years.
-Similar to Lulzec, there's a new group catching attention with different motivations, targeting larger organizations.
-Groups of this nature focus on data theft and extortion via means of social engineering - commonly, targeted spear phishing campaigns"
-story-blue: "The LAPSUS$ group emerged onto the scene a number of months ago, targeting high-profile organizations such as Nvidia, Samsung, and Ubisoft
-- making various demands that in some cases, resulted in either data dumps or screenshots of internal systems shared via the group's Telegram account.
-These were sometimes determined by user-voted polls within the group,
-suggesting that this is only the beginning of a series of attacks the group is undertaking more frequently as they gain press coverage."
-duration: 3
+#name: "Test scenario"
+#topic: finance
+#FQDN: myfinance.cu
+#story-red: "Financially motivated adversary groups executing ransomware attacks have rightfully gotten our attention in recent years.
+#Similar to Lulzec, there's a new group catching attention with different motivations, targeting larger organizations.
+#Groups of this nature focus on data theft and extortion via means of social engineering - commonly, targeted spear phishing campaigns"
+#story-blue: "The LAPSUS$ group emerged onto the scene a number of months ago, targeting high-profile organizations such as Nvidia, Samsung, and Ubisoft
+#- making various demands that in some cases, resulted in either data dumps or screenshots of internal systems shared via the group's Telegram account.
+#These were sometimes determined by user-voted polls within the group,
+#suggesting that this is only the beginning of a series of attacks the group is undertaking more frequently as they gain press coverage."
+#duration: 3
+
 networks:
   - name: vlan10
     tag: 10
@@ -16,8 +17,7 @@ networks:
     tag: 20
   - name: vlan30
     tag: 30
-hosts:ls
-
+hosts:
   - name: webserver
     type: docker
     networks: [ "20" ]
